feat(virtual): Virtual Drive PoC - Implement virtual drive that can be allocated and survives the self-test of `tcgsdiag`.
The virtual drive is added as an additional drive. The code from this patch allows the drive to be created by `core.NewCore()`.

The code can respond to some of the ComID requests but lacks a complex state machine to handle different requests.
The goal of the PoC is to implement what tcgsdiag needs to initialize and self-check the drive.

Signed-off-by: Christian Grönke <christian.groenke@9elements.com>